% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutations.R
\name{countTableSNV}
\alias{countTableSNV}
\title{The count table of each SNV, without context length}
\usage{
countTableSNV(mut_file = NULL, vcf_file = NULL, ref_genome = NULL)
}
\arguments{
\item{mut_file}{is the mutation vector obtained with getMutations.}

\item{vcf_file}{is the loaded VCF file containing the mutations. Needed if no mut_file is provided.}

\item{ref_genome}{is the reference genome. Needed if no mut_file is provided.}
}
\value{
Count table of the absolute frequency of each SNV, not considering the context length.
}
\description{
Get a count table of the absolute frequency of each SNV, not considering the context length.
}
\examples{
library(VariantAnnotation)
vcffile <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- readVcf(vcffile, "hg19")[1:20]
vcf <- renameSeqlevels(vcf, 'chr22')
library(BSgenome.Hsapiens.UCSC.hg19)
countTableSNV(vcf_file = vcf, ref_genome = Hsapiens)
mut_file <- getMutations(vcf, Hsapiens, 3)
countTableSNV(mut_file)
}
