% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/getMutations.R
\name{getMutations}
\alias{getMutations}
\title{Structured mutations from a VCF file}
\usage{
getMutations(vcf_file, ref_genome, context_length)
}
\arguments{
\item{vcf_file}{is the loaded VCF file containing the mutations}

\item{ref_genome}{is the reference genome}

\item{context_length}{is the length of the region displayed. It considers the mutation base and bases upstream and downstream of the mutation.}
}
\value{
Structured vector of single nucleotide variants
}
\description{
Get a structured vector of single nucleotide variants given a VCF file, a reference genome and a context length
}
\examples{
library(VariantAnnotation)
vcffile <- system.file("extdata", "chr22.vcf.gz", package="VariantAnnotation")
vcf <- readVcf(vcffile, "hg19")[1:20]
vcf <- renameSeqlevels(vcf, 'chr22')
library(BSgenome.Hsapiens.UCSC.hg19)
getMutations(vcf, Hsapiens, 5)
}
